Problems with scope
Made more additions /changes to the component and I seem to have figured out how to go about assigning the book data to a variable but I'm unsure as to how to use that data outside of the scope of the function itself. Trying to refer to the variable as anything other than a local instance of itself (this.example) causes a compiler error correction it back to this.